Backend/mcp/web_server/requirements.txt
mcp
langchain-community
aiohttp
beautifulsoup4
python-dotenv
